% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{as_numerical}
\alias{as_numerical}
\alias{as_categorical}
\alias{as_ordinal}
\title{Convert Variable Type}
\usage{
as_numerical(x, cols = NULL)

as_categorical(x, cols = NULL)

as_ordinal(x, cols = NULL, level = TRUE)
}
\arguments{
\item{x}{a data frame}

\item{cols}{columns names or numbers corresponding to variables to convert}

\item{level}{a boolean}
}
\value{
A data frame.
}
\description{
These functions convert variables (columns of a data frame) to a numeric
(\code{as_numerical()}), to a non-ordered categorical variable
(\code{as_categorical()}), or to an ordinal variable (\code{as_ordinal()}).
}
\section{Functions}{
\itemize{
\item \code{as_categorical}: Convert to categorical variable.

\item \code{as_ordinal}: Convert to ordinal variable.
}}

