% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{calc_dist}
\alias{calc_dist}
\title{Function to Compute Distance between Species}
\usage{
calc_dist(trait_df, trait_category_df, comb_name, classical_gower)
}
\arguments{
\item{trait_df}{a traits data frame with species as rows and traits as
columns}

\item{trait_category_df}{a data frame describing the categories of each
traits with 3 columns: trait name, trait category, and fuzzy category if
trait is fuzzy}

\item{comb_name}{a character vector of the traits combination to test}

\item{classical_gower}{if TRUE, do not use modified Gower distance from
Pavoine}
}
\value{
a distance matrix
}
\description{
For a given combination of traits, this function returns a distance matrix.
}
