% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_missing_trait_distance.R
\name{compute_missing_trait_distance}
\alias{compute_missing_trait_distance}
\title{Compute Traits Distance with Missing Traits}
\usage{
compute_missing_trait_distance(
  trait_df,
  trait_category_df,
  percent_missing,
  max_comb = 20,
  n_perm = 20,
  cores = 1,
  classical_gower,
  pos = 20
)
}
\arguments{
\item{trait_df}{a trait data.frame with species as rows and traits as columns}

\item{trait_category_df}{a data.frame describing the categories of each
traits with 3 columns: trait name, trait category and fuzzy category if
trait is fuzzy}

\item{percent_missing}{a numeric vector of percent of missing trait that
should be considered (should be between 0 and 1)}

\item{max_comb}{maximum number of combinations to sample}

\item{n_perm}{number of permutations for Mantel test}

\item{cores}{number of cores used in \code{mclapply()}}

\item{classical_gower}{if TRUE, do not use modified Gower from Pavoine}

\item{pos}{number of combinations to consider}
}
\value{
A data frame
}
\description{
Compute Traits Distance with Missing Traits
}
