% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depletion_plot.R
\name{depletion_plot}
\alias{depletion_plot}
\title{Influence of Each Trait and Sensibility of the Functional Space to a Specific Functional Trait}
\usage{
depletion_plot(data, version)
}
\arguments{
\item{data}{traits data frame output of \code{compute_missing_trait_distance()}}

\item{version}{types of graph (1, 2, or 3)}
}
\description{
Influence of Each Trait and Sensibility of the Functional Space to a Specific Functional Trait
}
