% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_plot.R
\name{dim_plot}
\alias{dim_plot}
\title{\strong{ADD DESCRIPTION}}
\usage{
dim_plot(data_miss, dim_pcoa)
}
\arguments{
\item{data_miss}{\strong{ADD DESCRIPTION}}

\item{dim_pcoa}{\strong{ADD DESCRIPTION}}
}
\description{
\strong{ADD DESCRIPTION}
}
