% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_simulation.R
\name{dim_simulation}
\alias{dim_simulation}
\title{Test Influence of Number of Traits and S on Simulated Data}
\usage{
dim_simulation(traits, tr_cat, threshold, cut, core)
}
\description{
Test Influence of Number of Traits and S on Simulated Data
}
