% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_funct.R
\name{dimension_funct}
\alias{dimension_funct}
\title{Compute the Number of Dimensions}
\usage{
dimension_funct(
  trait_df,
  dim_pcoa = 10,
  rep = 99,
  cores = 3,
  metric_scaled = TRUE,
  trait_category_df,
  classical_gower
)
}
\arguments{
\item{trait_df}{a traits data frame with species as rows and traits as
columns}

\item{dim_pcoa}{number of PCoA axes}

\item{rep}{number of replicates}

\item{cores}{number of cores (parallelization)}

\item{metric_scaled}{a boolean}

\item{trait_category_df}{a data frame describing the categories of each
traits with 3 columns: trait name, trait category, and fuzzy category if
trait is fuzzy}

\item{classical_gower}{if TRUE, do not use modified Gower distance from
Pavoine}
}
\description{
\strong{ADD DESCRIPTION}
}
