% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleton.R
\name{DPC}
\alias{DPC}
\title{Compute Singleton Clustering}
\usage{
DPC(
  X,
  metric = "euclidean",
  radius = "automatic",
  fraction = 0.01,
  mode = "automatic",
  rhomin = "automatic",
  deltamin = "automatic",
  density_filter = "saddle",
  rho_threshold = 0,
  removezeroes = FALSE
)
}
\arguments{
\item{metric}{\strong{ADD DESCRIPTION}}

\item{radius}{\strong{ADD DESCRIPTION}}

\item{fraction}{\strong{ADD DESCRIPTION}}

\item{mode}{\strong{ADD DESCRIPTION}}

\item{rhomin}{\strong{ADD DESCRIPTION}}

\item{deltamin}{\strong{ADD DESCRIPTION}}

\item{density_filter}{\strong{ADD DESCRIPTION}}

\item{rho_threshold}{\strong{ADD DESCRIPTION}}

\item{removezeroes}{\strong{ADD DESCRIPTION}}

\item{data}{\strong{ADD DESCRIPTION}}
}
\value{
\strong{ADD DESCRIPTION}
}
\description{
\strong{ADD DESCRIPTION}
}
