% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbow.R
\name{elbow}
\alias{elbow}
\title{Elbow Method to Detect Infection Point in a Concave Curve}
\usage{
elbow(data, xmin = NULL, xmax = NULL, ymin = NULL, ymax = NULL)
}
\arguments{
\item{data}{a two-columns data frame (x and y respectively).}

\item{xmin}{\link{optional} the min value on x-Axis (if missing in data).}

\item{xmax}{\link{optional} the max value on x-Axis (if missing in data).}

\item{ymin}{\link{optional} the min value on y-Axis (if missing in data).}

\item{ymax}{\link{optional} the max value on y-Axis (if missing in data).}
}
\value{
This function returns a 2-elements list with:
\itemize{
\item a numeric giving the value on x-Axis corresponding to the optimal
\item a data frame with the original values and two additional columns used in
the graphic.
}
}
\description{
This function implements the Elbow rule to detect infection point in a
concave curve.
}
