% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chose_n_combn.R
\name{get_combn}
\alias{get_combn}
\alias{chose_n_combn}
\title{Get Combinations with a Giving Number of Variables}
\usage{
get_combn(x, m, pos)

chose_n_combn(x, m, pos)
}
\arguments{
\item{x}{colnames of the matrix of traits}

\item{m}{number of modalities in the combination}

\item{pos}{number of possibilities considered}
}
\value{
\strong{WHAT?}
}
\description{
This function is an alternative to the default combn() and returns a subset
of combinaison.
}
\section{Functions}{
\itemize{
\item \code{chose_n_combn}: Batch execution of \code{get_combn()}
}}

