% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_miss_trait.R
\name{parse_miss_trait}
\alias{parse_miss_trait}
\title{Parse Name of Missing Trait Category}
\usage{
parse_miss_trait(miss_trait_name)
}
\arguments{
\item{miss_trait_name}{a character of type "misstrait_15\%_n4"}
}
\value{
Returns two numbers: the percentage of missing traits and the number
of traits
}
\description{
Parse Name of Missing Trait Category
}
