% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runifcor.R
\name{runifcor}
\alias{runifcor}
\title{Generate Correlated Traits Matrix Following Uniform Distribution}
\usage{
runifcor(x, rho)
}
\description{
Generate Correlated Traits Matrix Following Uniform Distribution
}
