% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_trait_combinations.R
\name{sample_trait_combinations}
\alias{sample_trait_combinations}
\title{Sample a Maximum Number of Combinations in All Combinations}
\usage{
sample_trait_combinations(combinations_list, max_n = 100)
}
\arguments{
\item{combinations_list}{a list of all combinations of traits}

\item{max_n}{maximum number of combinations to sample}
}
\value{
\strong{WHAT?}
}
\description{
Sample a Maximum Number of Combinations in All Combinations
}
