% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_results.R
\name{synth_results}
\alias{synth_results}
\title{Summarize Results}
\usage{
synth_results(trait_df, single_df, miss_final_df, dim_df)
}
\arguments{
\item{trait_df}{\strong{ADD DESCRIPTION}}

\item{single_df}{\strong{ADD DESCRIPTION}}

\item{miss_final_df}{\strong{ADD DESCRIPTION}}

\item{dim_df}{\strong{ADD DESCRIPTION}}
}
\value{
\strong{ADD DESCRIPTION}
}
\description{
\strong{ADD DESCRIPTION}
}
